feat(web): extract token usage view and ship public /tok pages
Publish stays in-app; visitors read /tok/@handle and /tok/leaderboard with the same charts.
