+++ apps/web/messages/en.json
@@ -633,6 +633,55 @@
           "capture": "Could not capture this page."
         }
       },
+      "publish": {
+        "trigger": "Publish",
+        "title": "Publish page",
+        "description": "Share a link on atmos.land. Updates overwrite the same page.",
+        "signInHint": "Sign in to publish a Token Usage page.",
+        "handleLabel": "Handle",
+        "handlePlaceholder": "builder",
+        "handleLocked": "Handles can only be claimed once.",
+        "visibility": {
+          "public": "Public",
+          "unlisted": "Unlisted"
+        },
+        "xPlaceholder": "username",
+        "githubPlaceholder": "username",
+        "includeCost": "Include estimated cost",
+        "copyLink": "Copy link",
+        "copied": "Copied",
+        "publish": "Publish",
+        "update": "Update",
+        "turnOff": "Turn off",
+        "rotateSecret": "New secret",
+        "unlistedNeedsSecret": "Unlisted links need the secret. Create a new secret to copy a working URL.",
+        "errors": {
+          "generic": "Could not publish this page.",
+          "usernameTaken": "That handle is taken.",
+          "handleLocked": "This handle cannot be changed.",
+          "invalidHandle": "Use 3–24 letters, numbers, or underscores.",
+          "invalidSocial": "Check the username and try again."
+        }
+      },
+      "publicPage": {
+        "notFound": "This page is not available.",
+        "generatedDate": "Generated {date}",
+        "generatedByPrefix": "Generated by ",
+        "generatedBySuffix": "",
+        "leaderboardLink": "Leaderboard"
+      },
+      "leaderboard": {
+        "updated": "Updated {time} · refreshes every 2 hours",
+        "tokensTitle": "Token",
+        "tokensEmpty": "No public pages yet.",
+        "costTitle": "Cost",
+        "costEmpty": "Nobody has published cost yet.",
+        "colRank": "Rank",
+        "colUser": "User",
+        "colX": "X",
+        "colGithub": "GitHub",
+        "outsideTop": "Your rank"
+      },
       "cookieAccess": {
         "title": "Browser cookies",
         "description": "Authorize cookie access to include Cursor and other signed-in web sessions.",
